requirements.txt
streamlit==1.29.0
requests==2.31.0
PyPDF2==3.0.1
google-generativeai==0.3.1